Extract types into lib
