+++ apps/tdsl/src/routes/shopping/list.tsx
@@ -1,10 +1,10 @@
+import { ShoppingListSchema } from '@tdsl/types';
 import { XCircle } from 'lucide-react';
 import { Fragment } from 'react';
 import { ActionFunctionArgs, json, redirect, useLoaderData } from 'react-router-dom';
 import * as z from 'zod';
 
 import { shoppingLists } from '../../data/shopping-lists';
-import { ShoppingListSchema } from '../../types';
 import { ActionList, ActionListItem, CreateInput, PageSection, SuspendedOutlet, VStack } from '../../ui';
 
 export const ShoppingLists: React.FC = () => {
