+++ libs/types/README.md
@@ -1,0 +1,3 @@
+# types
+
+This library was generated with [Nx](https://nx.dev).
